Made Player and Square regular classes instead of NamedTuples due to errors with value checking during instantiation. Added an attribute to the Player class containing its winning combination, e.g. "1111".
Added a function that checks for four-in-a-row on a given row by making a string representation of the row.
Added code that places a player's piece into a square and checks if there is a win in handle_move().
Added an attribute to the Logic class containing the winning coordinates, e.g. [(0, 2), (0, 3), ...].

Made the button colour constants private in the Graphics class.
Updated docstring formatting to be more like Google's style, and added the Attributes, Args, and Returns sections.