Imported pytest and wrote tests for `get_first_empty_square_in_column()` in a new file. Wrote a fixture for a Logic object.
Added __eq__() to Square for the test functions. Squares can now be compared.
Changed the function arguments of `is_valid_move()` and `handle_move()` such that they take a column index rather than a Square. This is simpler since only the column of the Square was needed, rather than the whole Square.
Decided not to change the `player_id` of the empty square retrieved in `handle_move()` since we only needed its coordinates.
Added a function to switch to the next player.

Updated the type annotation of functions with "-> None" if they return `None`.
Added pytest's cache files to .gitignore.
Updated the README such that it mentions pytest.