test_application.py: Renamed to test_logic.py since the tests only test the game's logic.
Renamed test functions to better describe their purpose and made their docstrings and comments more concise and easy to read.
Changed the column filling test function so that it only verifies that the column is full.
Changed the win detection test functions to make them easier to understand by limiting the play area to only what is needed for the tests.